+++ rollup.config.js
@@ -1,4 +1,6 @@
 
+import commonjs from 'rollup-plugin-commonjs';
+import resolve from 'rollup-plugin-node-resolve';
 const babel = require('rollup-plugin-babel');
 const pkg = require('./package');
 
@@ -20,6 +22,15 @@ export default {
     }
   ],
   plugins: [
+    commonjs({
+      namedExports: {
+        // left-hand side can be an absolute path, a path
+        // relative to the current directory, or the name
+        // of a module in node_modules
+        'node_modules/wcwidth/index.js': [ 'wcwidth' ]
+      }
+    }),
+    resolve(),
     babel(),
   ]
 };